+++ src/presentation/routes/context_route.py
@@ -3,14 +3,14 @@ from fastapi import APIRouter, Query, Request, Response, status
 
 from src.presentation.handler.context_handler import get_context_from_app
 from src.core.logs import debug, error
-
+from urllib.parse import unquote
 
 from ..handler.responses import MyResponse, NotEnoughQuestionsGeneratedError
 
 context_router = APIRouter()
 
 
-@context_router.get("/context/{item_name}/{input_identification}", response_model=MyResponse)
+@context_router.get("/context/{item_name}/{input_identification:path}", response_model=MyResponse)
 async def get_context(
     request: Request,
     item_name: str, 
@@ -18,9 +18,10 @@ async def get_context(
     response: Response,
     force_new_generation: bool = Query(False, description="Force new generation of questions, ignoring cached ones"),
     amount_question: int = Query(10, ge=1, le=20, description="Number of questions to generate"),
-    
+    selected_language: str = Query("English", description="Selected language for the questions")
     ):
 
+    input_identification = unquote(input_identification)
 
     user_uuid_id = request.headers.get("x-uuid")
 
@@ -30,7 +31,8 @@ async def get_context(
                 input_identification=input_identification, 
                 force_new_generation=force_new_generation,
                 amount_question=amount_question,
-                user_uuid_id=user_uuid_id
+                user_uuid_id=user_uuid_id,
+                selected_language=selected_language
         )
         if not context:
             response.status_code = status.HTTP_404_NOT_FOUND