+++ CHANGELOG.md
@@ -63,6 +63,7 @@ All notable changes to this project will be documented in this file.
 - *(pdf)* Fixed some errors to load correctly the quiz
 - *(cards)* Fixed the return from db or ai
 - *(quiz)* Catching the right keys dict questions
+- *(quiz)* Fixed query and context for quiz
 
 ### ⚙️ Miscellaneous Tasks